# backend/openapi.yaml
openapi: 3.1.0
info:
  title: StudySmart AI
  version: 0.1.0
paths:
  /auth/register:
    post:
      tags:
      - Authentication
      summary: Register
      operationId: register_auth_register_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/User'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /auth/login:
    post:
      tags:
      - Authentication
      summary: Login
      operationId: login_auth_login_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UserForm'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/courses/{course_id}:
    get:
      tags:
      - Courses
      summary: Get Course
      operationId: get_course_api_courses__course_id__get
      parameters:
      - name: course_id
        in: path
        required: true
        schema:
          type: integer
          title: Course Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/courses/:
    get:
      tags:
      - Courses
      summary: Get Courses
      operationId: get_courses_api_courses__get
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
    post:
      tags:
      - Courses
      summary: Create Course
      operationId: create_course_api_courses__post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CourseCreate'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/courses/instructor/{instructor_id}:
    get:
      tags:
      - Courses
      summary: Get Instructor Courses
      operationId: get_instructor_courses_api_courses_instructor__instructor_id__get
      parameters:
      - name: instructor_id
        in: path
        required: true
        schema:
          type: integer
          title: Instructor Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/assignments/:
    post:
      tags:
      - Assignments
      summary: Create Assignment
      operationId: create_assignment_api_assignments__post
      parameters:
      - name: week_id
        in: query
        required: true
        schema:
          type: integer
          title: Week Id
      - name: title
        in: query
        required: true
        schema:
          type: string
          title: Title
      - name: type
        in: query
        required: true
        schema:
          type: string
          title: Type
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/assignments/{assignment_id}:
    get:
      tags:
      - Assignments
      summary: Get Assignment
      operationId: get_assignment_api_assignments__assignment_id__get
      parameters:
      - name: assignment_id
        in: path
        required: true
        schema:
          type: integer
          title: Assignment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/assignments/week/{week_id}:
    get:
      tags:
      - Assignments
      summary: Get Week Assignments
      operationId: get_week_assignments_api_assignments_week__week_id__get
      parameters:
      - name: week_id
        in: path
        required: true
        schema:
          type: integer
          title: Week Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
  /api/submissions/:
    post:
      tags:
      - Submissions
      summary: Create Submission
      operationId: create_submission_api_submissions__post
      parameters:
      - name: assignment_id
        in: query
        required: true
        schema:
          type: integer
          title: Assignment Id
      - name: student_id
        in: query
        required: true
        schema:
          type: integer
          title: Student Id
      - name: content
        in: query
        required: true
        schema:
          type: string
          title: Content
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/submissions/grade:
    post:
      tags:
      - Submissions
      summary: Grade Submission
      operationId: grade_submission_api_submissions_grade_post
      parameters:
      - name: submission_id
        in: query
        required: true
        schema:
          type: integer
          title: Submission Id
      - name: grade
        in: query
        required: true
        schema:
          type: integer
          title: Grade
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/submissions/student/{student_id}:
    get:
      tags:
      - Submissions
      summary: Get Student Submissions
      operationId: get_student_submissions_api_submissions_student__student_id__get
      parameters:
      - name: student_id
        in: path
        required: true
        schema:
          type: integer
          title: Student Id
      - name: assignment_id
        in: query
        required: true
        schema:
          type: integer
          title: Assignment Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema: {}
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    CourseCreate:
      properties:
        title:
          type: string
          title: Title
        instructor_id:
          type: integer
          title: Instructor Id
      type: object
      required:
      - title
      - instructor_id
      title: CourseCreate
...
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    User:
      properties:
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        id:
          type: integer
          title: Id
      type: object
      required:
      - username
      - email
      - role
      - id
      title: User
    UserCreate:
      properties:
        username:
          type: string
          title: Username
        email:
          type: string
          title: Email
        role:
          type: string
          title: Role
        password:
          type: string
          title: Password
      type: object
      required:
      - username
      - email
      - role
      - password
      title: UserCreate
    UserForm:
      properties:
        email:
          type: string
          title: Email
        password:
          type: string
          title: Password
      type: object
      required:
      - email
      - password
      title: UserForm
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBasic:
      type: http
      scheme: basic